阳新实验室项目 1、退回重检调整完成

+++ yx-lab-backend/src/main/java/com/yx/lab/modules/sample/service/LabSampleService.java
@@ -108,7 +108,7 @@ public class LabSampleService {
         String prefix = DateUtil.format(new Date(), "yyyyMM");
         Long count = labSampleMapper.selectCount(new LambdaQueryWrapper<LabSample>()
                 .likeRight(LabSample::getSampleNo, prefix));
-        long next = count == null ? 1L : count.longValue() + 1L;
+        long next = count == null ? 1L : count + 1L;
         return prefix + String.format("%04d", next);
     }
